Moved 'common/mutex.h' into 'common/synchronization'. Renamed the 'Mutex' class to 'DirtyMutexClass'. Added a comment to dissuade people from using it. Moved 'Condition' into its own separate header file.
